nexchange/requirements.txt
djangorestframework==3.14.0
django
pytest==7.3.1
django-cors-headers
